+++ extensions-api/src/main/scala/pl/touk/nussknacker/engine/api/process/ClassExtractionSettings.scala
@@ -352,6 +352,7 @@ object ClassExtractionSettings {
       ),
       Set("canCastTo", "castTo")
     ),
+    MemberNamePredicate(ExactClassPredicate(classOf[java.lang.Object]), Set("canCastTo", "castTo")),
   )
 
   lazy val DefaultTypingFunctionRules: List[TypingFunctionRule] =