+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/spel/SpelExpressionSuggester.scala
@@ -197,6 +197,12 @@ class SpelExpressionSuggester(
               .queryEntriesByLabel(td.dictId, if (shouldInsertDummyVariable) "" else p.getName)
               .map(_.map(list => list.map(e => ExpressionSuggestion(e.label, td, fromClass = false, None, Nil))))
               .getOrElse(successfulNil)
+          case TypingResultWithContext(Unknown, staticContext) =>
+            Future.successful(
+              clssDefinitions.unknown
+                .map(c => filterClassMethods(c, p.getName, staticContext))
+                .getOrElse(Nil)
+            )
         }
         .getOrElse(successfulNil)
     }