+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/spel/typer/MethodReferenceTyper.scala
@@ -5,7 +5,6 @@ import cats.data.Validated.{Invalid, Valid}
 import pl.touk.nussknacker.engine.api.generics.ExpressionParseError
 import pl.touk.nussknacker.engine.api.typed.typing._
 import pl.touk.nussknacker.engine.definition.clazz.{ClassDefinition, ClassDefinitionSet, MethodDefinition}
-import pl.touk.nussknacker.engine.extension.CastTyping
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.IllegalOperationError.IllegalInvocationError
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.MissingObjectError.UnknownMethodError
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.{ArgumentTypeError, OverloadedFunctionError}
@@ -22,10 +21,8 @@ class MethodReferenceTyper(classDefinitionSet: ClassDefinitionSet, methodExecuti
       case TypedNull =>
         Left(IllegalInvocationError(TypedNull))
       case Unknown =>
-        if (CastTyping.isCastMethod(reference)) CastTyping.Instance.computeResultType(reference)
-        else if (methodExecutionForUnknownAllowed) Right(Unknown)
-        // todo: lbg suggestions?
-//        if (methodExecutionForUnknownAllowed) Right(Unknown)
+        // TODO: lbg Should we allow always access on Unknown to: canCastTo, castTo and toString?
+        if (methodExecutionForUnknownAllowed) typeFromClazzDefinitions(classDefinitionSet.unknown.toList)
         else Left(IllegalInvocationError(Unknown))
     }
   }