+++ docs/Changelog.md
@@ -45,7 +45,7 @@
 * [#6826](https://github.com/TouK/nussknacker/pull/6826) Security fix: added validation of expression used inside 
   indexer for Maps and Lists (for example `{1,2,3}[#otherList.remove(1) == null ? 0 : 0]`). This allowed executing 
   some types of unallowed expressions.
-* [#6840](https://github.com/TouK/nussknacker/pull/6840) Introduce extension functions in SpeL. Add canCastTo and castTo
+* [#6840](https://github.com/TouK/nussknacker/pull/6840) Introduce extension methods in SpeL. Add canCastTo and castTo
   extension functions.
 
 ## 1.17